# add two short test videos, use them in component view test