… pixeltable/tests/data/simple_videos/frame_count_black.mp4
… pixeltable/tests/data/simple_videos/frame_count_white.mp4