speeds component view test substantially.