@orm011 2024-03-25
I have now made a video that's the same length as the longest video previously used (462 frames) but with much smaller resolution. It's enough to make things go much faster. (all numbers are on a apple silicon v2 machine, so this is not about bad harware)

before:
44.42s call     pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_snapshot1

44.29s call     pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_snapshot2

36.39s call     pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_snapshot3

36.01s call     pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_snapshot4

32.15s call     pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_chained\_views

4.67s call     pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_add\_column

after:
slowest 10 durations ===============================================================================================

1.39s setup    pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_basic

0.95s call     pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_snapshot4

0.94s call     pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_snapshot2

0.92s call     pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_snapshot1

0.87s call     pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_snapshot3

0.81s call     pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_chained\_views

0.16s call     pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_basic

0.11s call     pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_add\_column

0.07s call     pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_update

0.03s setup    pixeltable/pixeltable/tests/test\_component\_view.py::TestComponentView::test\_snapshot1





<!-- Sent from Reviewable.io -->