@orm011 2024-03-25
*[`pixeltable/tests/utils.py` line 224 at r1](https://reviewable.io/reviews/pixeltable/pixeltable/73#gh-1468286510:-NtrIbKF0FsfF1G7FPaO:b1sgfnl) ([raw file](https://github.com/pixeltable/pixeltable/blob/54cd2e57ab3637166ed685ae3a06539884c0fb91/pixeltable/tests/utils.py#L224)):*
<details><summary><i>Previously, orm011 (Oscar Moll) wrote…</i></summary><blockquote>

I have now updated this to use a long video and a short video. The long video is as long as before. But the video is simpler. It's still much faster. Does that sound good?
</blockquote></details>

(simpler = low resolution frames)


<!-- Sent from Reviewable.io -->