@mkornacker 2024-03-25
I would like to take a completely different approach to solving this problem. Let's discuss.